bot/bot_requirements.txt
APScheduler==3.6.3
Pygments==2.11.2
aiofiles==0.6.0
aiohttp==3.8.1
aiosignal==1.2.0
asttokens==2.0.5
async-timeout==4.0.2
attrs==21.4.0
backcall==0.2.0
cachetools==4.2.2
certifi==2021.10.8
charset-normalizer==2.0.12
dataclass-factory==2.14
decorator==5.1.1
executing==0.8.3
frozenlist==1.3.0
idna==3.3
ipdb==0.13.9
ipython==8.2.0
jedi==0.18.1
matplotlib-inline==0.1.3
multidict==6.0.2
numpy==1.22.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
pip==22.0.4
prompt-toolkit==3.0.29
ptyprocess==0.7.0
pure-eval==0.2.2
pydub==0.24.1
python-telegram-bot==13.11
pytz-deprecation-shim==0.1.0.post0
pytz==2022.1
requests==2.27.1
setuptools==58.1.0
shazamio==0.1.1.0
six==1.16.0
stack-data==0.2.0
toml==0.10.2
tornado==6.1
traitlets==5.1.1
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.9
wcwidth==0.2.5
yarl==1.7.2
